.gitignore: Add IDE files
Adding these to ignore IntelliJ and VSCode metadata files.
We could consider having some of the IntelliJ files included
if were committed to having everyone on it, but we can come
back to it later.

Moving all of the .gitignores from the vite scaffold to the
top level, since we likely want these throughout if we
distribute the js across the codebase.