Prettier: Add JS/TS autoformatting
Autoformatting is great, since it prevents people from needing to nit
about various stylistic choices about the code in pull requests and
also reduces the amount of diffs that are not related to the logic of
the code and how it runs.

Using prettier here as it's a popular autoformatter.

If there are other options that people would prefer, we can edit the
settings and reformat the world when needed.