Mantine: Get mantine setup
The previous setup was with only vite + react.

Mantine will be the UI Component library for the project. Getting it
set up by adapting the code in the vite-mantine template.

https://github.com/mantinedev/vite-template
It includes more tools, such as stylelint and vitest.

The package manager was changed from yarn to pnpm.
Storybook integration was removed, but I would like to add it back.

I've left a TODO to figure out aliasing for postcss, as it doesn't seem
to read the alias from the .tsconfig. It still works with the relative
path, and only shows up in Welcome.test.tsx so it should be fine to
fix later.