Git: Add Aliases for pull and push
With the trunk based workflow, we are often rewriting git history.

`git pull origin main` is used to rebase changes in the local branch on
top of new changes on main.

`git push --force` was originally used to update the remotes of feature
branches, however, safer options with --force-with-lease and
--force-with-includes now exist.

Creating aliases for both of them to save time typing them.

Generally the rule of thumb is to run `git pullo` before every push.

The workflow is now:
1. Start from the main branch
2. `git pull` to update main
3. `git switch -c hchen/1` to make a new branch
4. Make changes and commit locally
5. `git pullo` to update local branch with new commits on main
6. Resolve merge conflicts in IntelliJ
7. `git push` to upload branch to a remote branch on Github
8. Create a pull request and get feedback
9. Update commits.
10. `git pushf` to force your changes onto Github.
11. Click the rebase and merge button in Github once approved.

L=PREP-13