Gerrit: Create script and documentation
Decided to get Gerrit set up, as I find it better for actually doing
code reviews. The only change to the local git workflow is to use
git up instead of git pushf.

Included all of the setup steps for both users and the project in case
I need to set up Gerrithub again.