Git: Add git-nb alias for creating a branch
`git nb branch/name` can now be used to create branches.

With the workflow we are using, it is more correct to use gerrit/main
or origin/main as the upstream for the branches.

Branches can be created while on other branches with this, but the
parent branch has to be created from main or its upstream manually
set to gerrit/main using:
`git branch branch/name --set-upstream-to=gerrit/main`

With this, we should likely just use gerrit/main as our source of truth.
We can set the main branch to track gerrit/main by default.

Unfortunately, there is no config for setting tracking to inherit, or
that would be used to make the git branch behavior do this by default.

TEST=manual
Created a new branch and checked its upstream using `git branch -vv`
Ran `git rebase` and it rebased correctly.
Ran `git status` and it checked the divergence correctly.

Change-Id: I5772dfdfce4da1221be23bac5b5dbc4bc6e36c67