Git: Ignore CI branches in Github
Adding a negative refspec to the config to prevent fetching CI branches
from Github.
https://stackoverflow.com/a/37165609

This method can also be used to ignore other people's personal branches.

TEST=manual
Created a branch in Github named `ci/test`
Fetching with this config does not fetch that branch
Fetching without this config fetches that branch

Change-Id: Idb62d90858081828f390571e0c1007b9ccc3f6c0