Personal: Refactor the pages folder into personal
The Home page in the js/src/pages folder was originally an example,
but it currently hosts the directory to go to the personal websites.

Moving it into personal/ and then renaming it, so that files are
consistent between Patina and Personal.

TEST=manual
Checked that the personal home page still worked.

Change-Id: I76a5ad87f91bf602db73175ddfeccbc347820ce5