CI/CD: Add gh auth check, add --quiet to git push
Added a section inside dependency checks of git up script that checks if
user is authenticated before running, exits and prompts user to run
gh auth login if not, proceeds if they are. Added a check so that if
verbose flag is off, will run git push with --quiet flag to not produce
any output unless an error occurs.

L=ENG-249
L=ENG-254
TEST=manual
- Ran "gh auth logout" on terminal and ran "git up" and script exited
- Ran "gh auth login" and ran "git up" and script executed
- Ran "git up" with -v flag=0 and git push ran with --quiet.

Change-Id: I5ea3d266b6f96e1a558216880da1aa4bfbbf480c