CI/CD: Add gh auth check, add --quiet to git push
Added a section inside dependency checks of git up script that checks if
user is authenticated before running, exits and prompts user to run
gh auth login if not, proceeds if they are. Changed verbose flag to be 0
by default. Added a check so that if verbose flag is present, will run
git push with --quiet flag to not produce output unless error occurs.

L=ENG-249
L=ENG-254
TEST=manual
- Ran "gh auth logout" on terminal and ran "git up" and script exited
- Ran "gh auth login" and ran "git up" and script executed
- Ran "git up" and "git up -v" and non-verbose and verbose outputs shown

Change-Id: I5ea3d266b6f96e1a558216880da1aa4bfbbf480c