PatNet: Add calendar functionality and highlight
Modified calendar to highlight dates that contain a certain event.
Created a function inside MainSection.tsx converting a date in raw
string ISO format into a Date object. Modified event cards to be able to
show the day highlighted in green and day of the week of an event by
creating an array of days of the week inside EventCards component.

L=ENG-270
L=ENG-277
TEST=manual
- Ran pnpm run dev locally.

Change-Id: I7e6df02875ff922dcd922e03a31db7ece0646a25