Lilac: Add Mantine RichTextEditors for title/desc
Added two Mantine RichTextEditor components to TitleDescEditor component
for Lilac. Added a CSS file for the component. The first text editor
is for the title of the issue and is locked to a bold font with no
controls like Linear. The second text editor is for the description of
the issue and allows bold, italic, strikethrough, link, quote, code
controls. Refactored Task directory name to lowercase task.

L=ENG-294
TEST=manual
- Ran pnpm run dev locally

Change-Id: I38dec7bd94016079f7c2b1dc0b1113f20c971778