Blog: Create authentication functionality
Created a LoginController with an endpoint /api/login that takes in a
raw string and compares it to an admin password. If comparison passes,
creates a new "auth" cookie in the browser.
Inside the LoginPage component, the login button now fetches to the
login endpoint and inputs the user inputted password string as the body
of the request.

L=ENG-350
TEST=manual
- gw bootRun on terminal, go to localhost:8080/login
- Inputted wrong password, nothing added to cookies
- Inputted correct password, "auth" cookie added to browser

Change-Id: I2c7f7d0d404700787f4905fdd20b9783521ef104