Navbar: Adjust mobile drawer
With the addition of more pages, they start forming two lines before
it changes to a burger menu. Bumping the breakpoint to sm.

Also moved the drawer close icon in and made it larger to more closely
align with the burger menu. It's still not perfectly on top, but it's
much closer.

TEST=manual

Change-Id: Ida6a9df46170a4c31b12019975b30798954f8735