Blogs: Enable Pagination
Added the active page to the listBlogReq. An offset is added to the
listBlogs query and is created using the active page and limit.
Some logic changes including resetting the active page to 1 when
changing the 'Blogs per page' to avoid indexing issues.

TEST=manual
Ran page locally and validated feature works

Change-Id: Id55767108d20319debb7437ea8acd3d05b8b58e0