ContentPage: Allow transition to consistent gap
Letting a user pass in Gap to the content page, so that we can
progressively transition the pages to have a consistent gap, and move
all of the layout stuff outside of the individual components.

Reducing the Hero module gap here, but we should do a pass on all of the
pages and remove it.

Having the gap be set inside ContentPage will allow us to handle mobile
more consistently, without having to add it everywhere ourselves.

TEST=manual

Change-Id: Iaee898f401dc610be7a2c942576eea1cc2cca139