PatChats: Implemented /join_patchats function
If the user isn't already in the database, an
instance for the user is created in the
patchats_member table as well as the discord_user
table with the Discord information.

GetDiscordUser now retrieves information based on
Discord ID instead of PatChats ID in order to
check if the user is in the database.

TEST=manual
Works in our test discord server.

Change-Id: Ib5c517b1933fae046a978f2e4dcacff4c6d4ca68