Revert "Patchats: create a client side op for broadcast"
This reverts commit 5a7b399444d3e7b6eff69bd5c56e64cb8f412387.

This commit breaks the build with a circular dependency.

Details
-------------------------------------------------------------

Description:

The dependencies of some of the beans in the application context form a cycle:

┌─────┐
|  discordUserClient defined in file [/Users/arklian/patina/build/classes/java/main/org/patinanetwork/discordbot/DiscordUserClient.class]
↑     ↓
|  JDAInitializer defined in file [/Users/arklian/patina/build/classes/java/main/org/patinanetwork/discordbot/JDAInitializer.class]
↑     ↓
|  JDAEventListener defined in file [/Users/arklian/patina/build/classes/java/main/org/patinanetwork/discordbot/JDAEventListener.class]
└─────┘

Action:

Relying upon circular references is discouraged and they are prohibited by default. Update your application to remove the dependency cycle between beans. As a last resort, it may be possible to break the cycle automatically by setting spring.main.allow-circular-references to true.

Change-Id: I18f6d8c7a695b9c28731b234bc4a4133eddc91c7