History: Created a history file
Used to track large deletions and refactors.
It could be used potentially for an exclusion file for git blame.

It's hard to know of files that no longer exist, so putting references
in a log here allows someone to go back to the last commit that those
files existed.

In this case, we're removing the personal websites from this repo,
but if any of the interns wanted to get their site, they can checkout
the git commit hash and grab the code from there.

TEST=none

Change-Id: I3e84e421ad3c4a73dbf492f07b351bb043de07c8