Docker: Use Dockerfile instead of Spring task.
Use a custom Dockerfile since we need to install dependencies for
Playwright to run.

The Docker image build is split up into two stages, build and runnable
in order to reduce image size.

We copy over the browsers and the Spring JAR from the build stage, but
we need to still install the libs onto the runnable image. These are
hard coded, since installing it through npx adds 700MB to the file size
and conflicts with the Java version of Playwright unless we keep the
versions pegged.

Updated the CD pipeline to call docker directly.

TEST=manual
Ran the container in Docker locally, and Playwright was working.
Hopefully the Git Action works, since it's not tested in this yet.

Change-Id: Ia8af272ca32054de67c9f581d3fd27b63b14cd95